Move spectral model under wsclean
