--- africanus/model/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["spectra"]
-
-from africanus.model.spec_model import spectra
